flask/flask_db/web/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6